Create base classes for Expression and Statement. Create reflection functions for Expression and Statement. Move class statements into main Statements.ts file to avoid circular dependency issues.
