+++ src/Scope.ts
@@ -6,8 +6,7 @@ import { XmlFile } from './files/XmlFile';
 import { CallableContainer, BsDiagnostic, FileReference } from './interfaces';
 import { Program } from './Program';
 import { BsClassValidator } from './validators/ClassValidator';
-import { NamespaceStatement, ParseMode, Statement, NewExpression, FunctionStatement } from './parser';
-import { ClassStatement } from './parser/ClassStatement';
+import { NamespaceStatement, ParseMode, Statement, NewExpression, FunctionStatement, ClassStatement } from './parser';
 import { standardizePath as s, util } from './util';
 import { globalCallableMap } from './globalCallables';
 import { FunctionType } from './types/FunctionType';
